requirements.txt
ultralytics
roboflow
streamlit
opencv-python
Pillow